Fix bf16-vs-fp32 dtype crash at joint_epoch in slmadv
When mixed_precision='bf16', s_trg and ref are produced inside the
autocast block earlier in the iter and end up as bf16 tensors. The
slmadv call site is outside any autocast block, and slmadv's internal
forwards (predictor -> text_encoder -> AdaIN1d.fc) hit Linear layers
with fp32 weights, raising:

    RuntimeError: mat1 and mat2 must have the same dtype,
                  but got BFloat16 and Float

triggered when epoch first reaches joint_epoch.

Cast s_trg.detach() and ref to fp32 at the slmadv call site so its
internal model forwards see fp32 inputs that match the fp32 weights.
This keeps slmadv decoupled from AMP (matching its current design,
which has no internal autocast wrappers) and is a no-op when AMP is
disabled. Applied in train_finetune.py and train_second.py.

+++ train_second.py
@@ -568,13 +568,17 @@ def main(config_path):
                         ref_lengths = input_lengths
                         ref_texts = texts
 
+                    # slmadv is not wrapped in autocast; its internal model
+                    # forwards run in fp32, so style inputs must be cast back
+                    # from bf16 to match the fp32 Linear weights.
                     slm_out = slmadv(i,
                                      y_rec_gt,
                                      y_rec_gt_pred,
                                      waves,
                                      mel_input_length,
                                      ref_texts,
-                                     ref_lengths, use_ind, s_trg.detach(), ref if multispeaker else None)
+                                     ref_lengths, use_ind, s_trg.detach().float(),
+                                     ref.float() if multispeaker else None)
 
                     if slm_out is None:
                         continue
